BR-25 merge dev branch
